requirements.txt
# requirements.txt
Flask>=2.0
transformers>=4.20.0 # Or a version compatible with your models
torch>=1.9.0 # Or a version compatible with your system (CPU or CUDA)
PyMuPDF>=1.18.0
Pillow>=8.0.0